Self Portrait Fix
The reason for this sites existence is to show the data of how a funky goblin lives. In order to do this, I went on a magical quest out into the world, logging the daily activities that I do that make me feel the most funky goblinist. When planning out this website, I didn't have a plan. I just let things build on themselves naturally. I had a few ideas from the start like the down-here being rotated with different color letters being shown when highlighted over. After that I just started playing around with the hover feature a lot and figured out how to have it make text move. I will say, this self portrait helped me in understanding how to use the scroll feature and positioning features a bit better. I also learned the importance of classes in HTML/CSS